more error checking

+++ src/AFRS/Requests/DeleteRequest.java
@@ -15,10 +15,14 @@ public class DeleteRequest implements Request {
     @Override
     public ArrayList<String> doRequest(String[] params) {
         //format should passenger,origin,destination
+        ArrayList<String> response = new ArrayList<>();
+        if(params.length != 3) {
+            response.add("error,reservation not found");
+            return response;
+        }
         String passenger = params[0];
         String origin = params[1];
         String destination = params[2];
-        ArrayList<String> response = new ArrayList<>();
         response.add(reservationDB.delete(passenger, origin, destination));
 
         return response;
